requirements.txt
streamlit
opencv-python-headless
pillow
pytesseract
pdf2image
scikit-learn
matplotlib